+++ CHANGES.rst
@@ -4,10 +4,11 @@ Changelog
 Here you can see the full list of changes between each SQLAlchemy-Utils release.
 
 
-0.27.12 (2014-12-xx)
+0.28.0 (2014-12-12)
 ^^^^^^^^^^^^^^^^^^^^
 
 - Fixed PhoneNumber string coercion (#93)
+- Added observes decorator (generates decorator will be deprecated later)
 
 
 0.27.11 (2014-12-06)